add patient

+++ Services/PatientRepository.cs
@@ -11,8 +11,6 @@ namespace DoAn_API.Services
             _context = context;
         }
 
-
-
         public PatientVM Add(PatientVM patient)
         {
             var newPatient = new Patient
@@ -42,13 +40,16 @@ namespace DoAn_API.Services
             };
         }
 
-
-
         public void Delete(int id)
         {
             var patient = _context.Patients.SingleOrDefault(patient => patient.userId == id);
             if (patient != null)
             {
+                if (!IsPatientUser(patient))
+                {
+                    return;
+                };
+
                 _context.Patients.Remove(patient);
                 _context.SaveChanges();
             }
@@ -65,6 +66,11 @@ namespace DoAn_API.Services
             var patient = _context.Patients.SingleOrDefault(patient => patient.userId == id);
             if (patient != null)
             {
+                //if (!IsPatientUser(patient))
+                //{
+                //    return null;
+                //};
+
                 return new PatientVM
                 {
                     userId = patient.userId,
@@ -84,13 +90,17 @@ namespace DoAn_API.Services
             }
         }
 
-
-
         public void Update(PatientVM patient)
         {
             var type = _context.Patients.SingleOrDefault(patient => patient.userId == patient.userId);
+
             if (type != null)
             {
+                if (!IsPatientUser(type))
+                {
+                    return;
+                };
+
                 type.email = patient.email;
                 type.password = patient.password;
                 type.image = patient.image;
@@ -101,6 +111,31 @@ namespace DoAn_API.Services
                 type.dob = patient.dob;
                 _context.SaveChanges();
             }
+        }
+
+        public async Task CreatePatient(Patient patient)
+        {
+            // EF Core sẽ tự động thêm vào bảng Users trước, sau đó thêm vào bảng Patient
+            await _context.Patients.AddAsync(patient);
+            await _context.SaveChangesAsync();
+        }
+
+        public bool IsPatientUser(Patient patient)
+        {
+            if (patient.roles == null)
+            {
+                return false;
+            }
+
+            foreach (var role in patient.roles)
+            {
+                if (role.roleId != 1)
+                {
+                    return false;
+                }
+            }
+
+            return true;
         }
     }
 }
